# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trae Platform API
  description: |
    设备与服务器测试运维平台 API 文档。
    包含仪表盘、一键任务、Monkey 测试、性能监控、日志监控、点歌、ARM 压测等模块。
  version: 1.0.0
  contact:
    name: Trae Platform
servers:
  - url: /
    description: 相对路径（部署时替换为实际域名）
...
paths:
...
  /unified/api/reports:
    get:
      tags: [一键任务]
      summary: 报告列表（可筛选）
      parameters:
        - name: module
          in: query
          schema: { type: string }
        - name: kind
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: keyword
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: reports 数组
...
  /server_stress/api/stress/start:
    post:
      tags: [ARM 压测]
      summary: 启动压测
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                server_id: { type: string }
                cpu_cores: { type: integer }
                cpu_load: { type: integer }
                timeout: { type: integer }
      responses:
        '200':
          description: success, message
...